Fix FreeBSD compiler warnings.

+++ tests/libntp/timevalops.c
@@ -76,7 +76,10 @@ static bool AssertTimevalClose(const struct timeval m, const struct timeval n,
 		return true;
 	else 
 	{
-		printf("m_expr which is %ld.%lu \nand\nn_expr which is %ld.%lu\nare not close; diff=%ld.%luusec\n", m.tv_sec, m.tv_usec, n.tv_sec, n.tv_usec, diff.tv_sec, diff.tv_usec); 
+		printf("m_expr which is %ld.%lu \nand\nn_expr which is %ld.%lu\nare not close; diff=%ld.%luusec\n",
+		       (long)m.tv_sec, m.tv_usec,
+		       (long)n.tv_sec, n.tv_usec,
+		       (long)diff.tv_sec, diff.tv_usec); 
 		//I don't have variables m_expr and n_expr in unity, those are command line arguments which only getst has!!!
 
 		return false;
